platforms: Ensure proper pcr prediction for azure
Adds test-fixtures for azure, and makes sure that tests for that
platform are run in github ci.

Adds the host-platform matrix variable to the tests job. For now, we
will only test the azure platform with the fedora coreos 42 image as
that's the only data we've gathered for now.

It also excludes secureboot-disabled tests for the azure platform, as we
haven't measured that case for now either.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>