uefi::secureboot: Delete module, it's not needed
collect_secure_boot_hashes is not longer used. That makes all the fancy
stuff written around it as Traits and such not longer needed too. Which
is a good thing as, honestly, it added unneeded complexity to the
source.

To remove it cleanly, this commit moves the secureboot_db implementation
for EFIVarsLoader from the SecureBootdbLoader (now removed) trait into
the actual context for method implementations of EFIVarsLoader.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>