tree: Simple event tree implementation
Now simple trees can be built, it can be checked if nodes nodes are
leafs or root nodes and tree branches can be computed.  It's a generic
tree that can hold any type T. All of this is unit-tested.

Children capacity of each node is set to 2 for now, assuming we're
processing images in pairs.

It also creates, a type alias called ResultNode. This is basically a
Node that holds a Result<T, E>. Then implements the following methods:
    - new_ok; an alias for Node::new(OK(data<T>))
    - new_err; an alias for Node::new(Err(error<E>))
    - valid_branches: Function that returns the branches that do not
      hold any error. Even if there is just one, that branch get's
      filtered out.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>