tpmevents.combine: Combine Vec<TPMEvent>s
Combine Vec<TPMEvent> from image "this" and image "that" into a vector
of all possible solutions.

This is done by constructing a tree of EventNodes. EventNodes are tree
nodes that hold results. Results can be Ok(event) for cases in which the
event could be picked from one image or the other safely, or Err(error)
for cases in which an event group conflict was hit.

A mechanism to recover from those errors is not needed at the current
stage, to it's been decided to safely drop those errors during Pcr
compilation.

Also support combining N images by adding a wrapper that goes through
all possible image combinations. It's not the optimal way, but it leads
to code that is easier to read and handle, where performance is not key
(compared to what it usually takes to pull a bootable image).

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>